Animations and more about me info also I changed the readme again.
I updated the animations on the buttons to transition back instead of snapping.

The about page on the website got a little bit more information, I also just recently did something with the google search console or something like that.

I can't seem to get favicons to work... I'm not sure why.